requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.12.0
    # via -r requirements.in
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   httpcore
    #   starlette
asyncpg==0.28.0
    # via -r requirements.in
build==1.0.3
    # via pip-tools
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
click==8.1.7
    # via
    #   pip-tools
    #   uvicorn
deprecated==1.2.14
    # via limits
fastapi==0.103.1
    # via -r requirements.in
flake8==6.1.0
    # via -r requirements.in
greenlet==2.0.2
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.18.0
    # via httpx
httpx==0.25.0
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   httpx
importlib-resources==6.1.0
    # via limits
iniconfig==2.0.0
    # via pytest
limits==3.6.0
    # via slowapi
loguru==0.7.2
    # via -r requirements.in
mako==1.2.4
    # via alembic
markupsafe==2.1.3
    # via mako
mccabe==0.7.0
    # via flake8
packaging==23.1
    # via
    #   build
    #   limits
    #   pytest
pip-tools==7.3.0
    # via -r requirements.in
pluggy==1.3.0
    # via pytest
pycodestyle==2.11.0
    # via flake8
pycryptodome==3.19.0
    # via -r requirements.in
pydantic==2.3.0
    # via fastapi
pydantic-core==2.6.3
    # via pydantic
pyflakes==3.1.0
    # via flake8
pyproject-hooks==1.0.0
    # via build
pytest==7.4.2
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==0.21.1
    # via -r requirements.in
slowapi==0.1.8
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==2.0.21
    # via
    #   -r requirements.in
    #   alembic
starlette==0.27.0
    # via fastapi
typing-extensions==4.8.0
    # via
    #   alembic
    #   fastapi
    #   limits
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
uvicorn==0.23.2
    # via -r requirements.in
wheel==0.41.2
    # via pip-tools
wrapt==1.15.0
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
